Handle downloads of Reddit videos without audio

+++ rvidmaker/videos/reddit.py
@@ -3,6 +3,7 @@
 import ffmpeg
 import os
 import requests
+import shutil
 
 from rvidmaker.utils import get_random_path
 from .interface import DownloadException, VideoRef
@@ -16,7 +17,14 @@ if not os.path.exists(_TEMP_DOWNLOAD_DIR):
 class RedditVideoRef(VideoRef):
     """References videos hosted on Reddit"""
 
-    def __init__(self, title, author, video_url, audio_url):
+    def __init__(self, title, author, video_url, audio_url=None):
+        """
+        Args:
+            title (str): Title of the video.
+            author (str): Author of the video.
+            video_url (str): Remote URL for video.
+            audio_url (str): Remote URL for audio. `None` if there is no audio.
+        """
         self.title = title
         self.author = author
         self.video_url = video_url
@@ -30,30 +38,46 @@ class RedditVideoRef(VideoRef):
 
         # Download video and audio to temporary files.
         temp_video_path = get_random_path(_TEMP_DOWNLOAD_DIR, "mp4")
-        temp_audio_path = get_random_path(_TEMP_DOWNLOAD_DIR, "mp4")
         # TODO: Download video and audio asynchronously
         with open(temp_video_path, "wb") as f:
             req = requests.get(self.video_url)
-            # TODO: Check `req.status_code`
+            if req.status_code != 200:
+                raise DownloadException(
+                    "Failed to download video from {}: {} response".format(
+                        self.video_url, req.status_code
+                    )
+                )
             f.write(req.content)
-        with open(temp_audio_path, "wb") as f:
-            req = requests.get(self.audio_url)
-            # TODO: Check `req.status_code`
-            f.write(req.content)
+        if self.audio_url is not None:
+            temp_audio_path = get_random_path(_TEMP_DOWNLOAD_DIR, "mp4")
+            with open(temp_audio_path, "wb") as f:
+                req = requests.get(self.audio_url)
+                if req.status_code != 200:
+                    raise DownloadException(
+                        "Failed to download video from {}: {} response".format(
+                            self.audio_url, req.status_code
+                        )
+                    )
+                f.write(req.content)
 
-        # Combine video and audio
-        video = ffmpeg.input(temp_video_path)
-        audio = ffmpeg.input(temp_audio_path)
-        try:
-            ffmpeg.concat(video, audio, v=1, a=1).output(output_path).run(
-                quiet=True, overwrite_output=True
-            )
-        except ffmpeg.Error:
-            raise DownloadException("Failed to combine video and audio with FFmpeg")
-
-        # Delete temporary files
-        os.remove(temp_video_path)
-        os.remove(temp_audio_path)
+            # Combine video and audio
+            video = ffmpeg.input(temp_video_path)
+            audio = ffmpeg.input(temp_audio_path)
+            try:
+                ffmpeg.concat(video, audio, v=1, a=1).output(output_path).run(
+                    quiet=True, overwrite_output=True
+                )
+            except ffmpeg.Error:
+                if os.path.exists(output_path):
+                    os.remove(output_path)
+                print("ERROR:", output_path)  # DEBUG
+                raise DownloadException("Failed to combine video and audio with FFmpeg")
+            finally:
+                # Delete temporary files
+                os.remove(temp_video_path)
+                os.remove(temp_audio_path)
+        else:
+            shutil.move(temp_video_path, output_path)
 
         return output_path
